fix: Adding a readOnlyFlag
The is an issue where when not doing offline first requesting a record that doesn't exist would result in a new empty record being created (locally) with data of zero. This would break bindings for some people. Instead what we do now is use a recordReadOnlyMode flag which indicates that we expected most records to be only ever read from server, and hence will only be ready on data load. You can explicitly mention the things you do want to use via a recordPrefixWriteWhitelist.

Also added a state from unsubscribing to deleted, because so many states.